Require TLS certificate in TLS authentication tests
Otherwise we get platform-specific failures and even generic client exceptions.
This case is potentially valid but is likely to be really rare (if you use TLS,
you probably also authenticate, using credentials or the certificate).

In the original patch this scenario was expected to fail. We now go back
to square 1.