Check ClientID at connect
Provide the possibilty for advance authentication mechanisms (such as via the rabbitmq http auth plugin) to check the clientId, and enforce specific rules.

This enhancement used in conjuction with the rabbitmq http auth plugin allows to pass the client_id as an additionnal field when querying the user_path url, and perform specific validation of the clientId vs. username as part of the authentication/authorization process.